Adds Meeting page and details page
Add a view that returns all the meetings:

Make a template that displays all the meeting items in a list

modify url.py to contain this view. 

Add a link to the base.html to add meetings to the menu

Try the link. Troubleshoot until it works

Make a view that takes an int id as an parameter and retrieves the meeting that has that id

modify the meeting template so that the title is a link to a detail page and each link returns the id

Create the template for the detail

modify the url to include the id

Try the link and troubleshoot until it works

Upload the code to GitHub and post the link for the assignment